% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunc.R
\name{estimate_beta}
\alias{estimate_beta}
\title{OLS estimator}
\usage{
estimate_beta(y, X)
}
\arguments{
\item{y}{A numerical vector, outcomes.}

\item{X}{Also a numerical vector, observations.}
}
\value{
A numerical vector of (X^T X)^{-1} X^T y.
}
\description{
OLS estimator
}
