% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunc.R
\name{my_theme}
\alias{my_theme}
\title{A custom ggplot2 theme with no grid lines, a light gray background, and bold axis titles.}
\usage{
my_theme()
}
\value{
A ggplot2 theme object.
}
\description{
A custom ggplot2 theme with no grid lines, a light gray background, and bold axis titles.
}
